Rename ComponentSortBottomSheet and ComponentSortViewModel test files to reflect new package structure and update references to ComponentSortBottomSheetUI
